*Final Release commit *Fix bugs *Improvements *Add Paging support to databases *Refactor classes and moving them to appropriate packages *Fix bug of not showing more button in notification list *Add Terms and Conditions, Privacy Policy and Website links *Moved every Repository and utils to use Singleton coding pattern to see some improvement *Create new logic for updating overall attendance while individual attendance is changed *New logic elements the use of ViewModelFactory, OverallAttendanceForSubjectViewModel and respective repository *Fix some typos *Fix problem of multiple touches of checkbox in notification preference *Use lambdas instead at many places to keep code clean and short *Disable Auto Attendance Feature *Add new layout to Notes *Add new functionality to Notes Layout such as delete, edit from notes directly *Remove unnecessary things

+++ app/src/main/java/com/vyas/pranav/studentcompanion/ui/fragments/NotesListFragment.java
@@ -9,15 +9,18 @@ import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProviders;
+import androidx.paging.PagedList;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 import com.vyas.pranav.studentcompanion.R;
 import com.vyas.pranav.studentcompanion.adapters.NoteRecyclerAdapter;
 import com.vyas.pranav.studentcompanion.data.maindatabase.MainDatabase;
-import com.vyas.pranav.studentcompanion.data.models.NotesEntry;
+import com.vyas.pranav.studentcompanion.data.notedatabase.NotesEntry;
 import com.vyas.pranav.studentcompanion.utils.AppExecutors;
 import com.vyas.pranav.studentcompanion.viewmodels.NoteViewModel;
 import com.vyas.pranav.studentcompanion.viewmodels.NotesViewModelFactory;
@@ -37,6 +40,11 @@ public class NotesListFragment extends Fragment implements NoteRecyclerAdapter.O
     @BindView(R.id.rv_notes_list_fragment_list)
     RecyclerView rvNotesList;
 
+    @BindView(R.id.placeholder_notes)
+    ConstraintLayout placeHolder;
+
+    private NoteRecyclerAdapter mAdapter;
+
     public NotesListFragment() {
     }
 
@@ -51,26 +59,54 @@ public class NotesListFragment extends Fragment implements NoteRecyclerAdapter.O
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        NoteRecyclerAdapter mAdapter = new NoteRecyclerAdapter();
+        mAdapter = new NoteRecyclerAdapter();
         mAdapter.setOnNotesDeleteClickedListener(this);
         rvNotesList.setAdapter(mAdapter);
-        rvNotesList.setLayoutManager(new LinearLayoutManager(getContext(), RecyclerView.VERTICAL, false));
+        LinearLayoutManager llm = new LinearLayoutManager(getContext(), RecyclerView.VERTICAL, false);
+        rvNotesList.setLayoutManager(llm);
         if (getArguments().getInt(EXTRA_NOTES_DISPLAY_TYPE, TYPE_ALL_NOTES_SHOW) == TYPE_TILL_TODAY_NOTES_SHOW) {
             long dateInMillis = new Date().getTime() - TimeUnit.DAYS.toMillis(1);
             NotesViewModelFactory factory = new NotesViewModelFactory(new Date(dateInMillis), getContext());
             NotesViewModelForDate viewModelForDate = ViewModelProviders.of(getActivity(), factory).get(NotesViewModelForDate.class);
-            viewModelForDate.getNotes().observe(this, mAdapter::submitList);
+            viewModelForDate.getNotes().observe(this, this::submit);
         } else {
             NoteViewModel viewModel = ViewModelProviders.of(getActivity()).get(NoteViewModel.class);
-            viewModel.getAllNotes().observe(this, mAdapter::submitList);
+            viewModel.getAllNotes().observe(this, this::submit);
+        }
+    }
+
+    private void submit(PagedList<NotesEntry> notesEntryList) {
+        if (notesEntryList.size() > 0) {
+            showPlaceHolder(false);
+            mAdapter.submitList(notesEntryList);
+        } else {
+            showPlaceHolder(true);
+        }
+    }
+
+    private void showPlaceHolder(boolean isShown) {
+        if (isShown) {
+            rvNotesList.setVisibility(View.GONE);
+            placeHolder.setVisibility(View.VISIBLE);
+        } else {
+            rvNotesList.setVisibility(View.VISIBLE);
+            placeHolder.setVisibility(View.GONE);
         }
     }
 
     @Override
     public void OnNotesDeleteClicked(NotesEntry note) {
-        AppExecutors.getInstance().diskIO().execute(() -> {
-            MainDatabase.getInstance(getContext()).noteDao().deleteNote(note);
-            AppExecutors.getInstance().mainThread().execute(() -> Toast.makeText(getContext(), "Note was deleted successfully!", Toast.LENGTH_SHORT).show());
-        });
+        new MaterialAlertDialogBuilder(getContext())
+                .setMessage("Do you want to delete note?\nYou can not undo action afterwards.")
+                .setTitle(R.string.title_attention)
+                .setPositiveButton("Delete", (dialogInterface, i) -> {
+                    AppExecutors.getInstance().diskIO().execute(() -> {
+                        MainDatabase.getInstance(getContext()).noteDao().deleteNote(note);
+                        AppExecutors.getInstance().mainThread().execute(() -> Toast.makeText(getContext(), "Note was deleted successfully!", Toast.LENGTH_SHORT).show());
+                    });
+                }).setNegativeButton("No", (dialogInterface, i) -> {
+            dialogInterface.dismiss();
+        }).show();
+
     }
 }
